Rename about page to search, correct env folder name
