--- StruggleBus/src/pages/about/about.scss
@@ -1,3 +1,0 @@
-page-about {
-
-}
