+++ StruggleBus/src/pages/search/search.html
@@ -1,6 +1,6 @@
 <ion-header>
-  <ion-navbar class="nav-title">
-    <ion-title text-center>
+  <ion-navbar class="nav-title" align-title="center">
+    <ion-title>
       <p class="nav-title">{{ quarter | quarterPipe }}</p>
     </ion-title>
   </ion-navbar>
