+++ StruggleBus/src/pages/search/search.scss
@@ -1,0 +1,3 @@
+page-search {
+
+}
